+++ Lights/PointLight.cs
@@ -9,14 +9,14 @@ namespace Dull.Lights
     internal class PointLight : ILight
     {
         private Vector3 _position;
-        private int _intensity;
+        private float _intensity;
         private Vector3 _color;
 
         private static int _objectSize = 2;//in Vector4
         private LightType _type = LightType.PointLight;
         Vector4[] _data;
         private int _byteOffset = -1;
-        public PointLight(Vector3 position, int intensity, Vector3 color)
+        public PointLight(Vector3 position, float intensity, Vector3 color)
         {
             _position = position;
             _intensity = intensity;
@@ -24,7 +24,7 @@ namespace Dull.Lights
             _data = new Vector4[_objectSize];
             SetStd140Data();
         }
-        public PointLight(Vector3 position, int intensity, Vector3i color)
+        public PointLight(Vector3 position, float intensity, Vector3i color)
         {
             _position = position;
             _intensity = intensity;
@@ -93,12 +93,12 @@ namespace Dull.Lights
             _color = color;
         }
 
-        public int GetIntensity()
+        public float GetIntensity()
         {
             return _intensity;
         }
 
-        public void SetIntensity(int intensity)
+        public void SetIntensity(float intensity)
         {
             _intensity = intensity;
         }
